Resolve identifiers to their local declaration, ignore missing
We now make an attempt to find the local declaration for every
Identifier node in the local module's AST.

Due to the way some JS toolchains work, some Identifiers may have
a value determined at build time or even at run time.

Because we emphatically do not want to deal with the various JS
toolchains, we just pretend unknown Identifiers refer to some
undefined-like object whose properties are recursively undefined-like.

And we happily inject that magical object in the global context
to let V8 evaluate the module as if all were right in the world.