Parse and support Flow syntax
Flow support has been added to our AST, the types are exposed
as regular identifiers and resolved normally.

We've replaced Babel's transform with just a Babylon parse(),
and we now do the stripping of flow types ourselves.
We need to strip flow types for v8 either way, but doing it ourselves
has the advantage of only minimally processing the source input,
and preserving some of the location information in the AST.

We can now rely on the fact that a node in the AST can be mapped
to its position in the original and transformed source code
(Note that line and column numbers are not currently preserved
in the transformed code. This is fine as users shouldn't see them).