Import AST directly from V8 objects, instead of going through JSON
We used to serialize the V8 objects to json, then deserialize that
to a C++ JSON tree, and then import that into a C++ AST.

That's a pretty clear performance improvement. Now we're still doing
the whole "parse Javascript using a Javascript parser" thing,
but at least everything around that is going to be fast!