Suggest possible missing awaits, off by default, with false positives
This is only a suggestion, not a warning, because it's very simple
and WILL cause false positives.
We currently consider any direct call to an async function to be
missing an await, unless we're just returning the result through.

Assigning the promise to a variable, intentionally starting async
functions in the background without keeping a promise, or returning
a promise from a non-async function all trigger false positives.